Feature/macos build qemu

+++ sys/kernel/syscall.c
@@ -179,7 +179,7 @@ int sysCheckPid(int pid, int *ptr) {
  ************************************************************************/
 int sysGetFreePage(struct thread *td, uint32_t *count) {
 
-  td->td_retval[0] = vmm_getFreeVirtualPage(_current->id, *count, VM_THRD);
+  td->td_retval[0] = (int)vmm_getFreeVirtualPage(_current->id, *count, VM_THRD);
   return(0);
   //return(vmm_getFreeVirtualPage(_current->id, *count, VM_TASK));
 }
